requirements.txt
scipy==1.10.1
sympy==1.12.1
numpy==1.24.4
matplotlib==3.7.5
opencv_python==4.9.0.80
torch==1.12.1
torchvision==0.13.1
tensorboard==2.14.0
tqdm==4.66.4
Pillow==10.3.0
h5py==3.11.0